Adding ShapeVector to file list
